+++ mne/viz/tests/test_topomap.py
@@ -22,7 +22,7 @@ from mne.time_frequency.tfr import AverageTFR
 from mne.utils import slow_test
 
 from mne.viz import plot_evoked_topomap, plot_projs_topomap
-from mne.viz.topomap import _check_outlines
+from mne.viz.topomap import _check_outlines, _onselect
 
 # Set our plotters to test mode
 import matplotlib
@@ -171,6 +171,7 @@ def test_plot_topomap():
 def test_plot_tfr_topomap():
     """Test plotting of TFR data
     """
+    import matplotlib as mpl
     import matplotlib.pyplot as plt
     raw = _get_raw()
     times = np.linspace(-0.1, 0.1, 200)
@@ -181,6 +182,19 @@ def test_plot_tfr_topomap():
     tfr = AverageTFR(raw.info, data, times, np.arange(n_freqs), nave)
     tfr.plot_topomap(ch_type='mag', tmin=0.05, tmax=0.150, fmin=0, fmax=10,
                      res=16)
+
+    eclick = mpl.backend_bases.MouseEvent('button_press_event',
+                                          plt.gcf().canvas, 0, 0, 1)
+    eclick.xdata = 0.1
+    eclick.ydata = 0.1
+    eclick.inaxes = plt.gca()
+    erelease = mpl.backend_bases.MouseEvent('button_release_event',
+                                            plt.gcf().canvas, 0.9, 0.9, 1)
+    erelease.xdata = 0.3
+    erelease.ydata = 0.2
+    pos = [[0.11, 0.11], [0.25, 0.5], [0.0, 0.2], [0.2, 0.39]]
+    _onselect(eclick, erelease, tfr, pos, 'mag', 1, 3, 1, 3, 'RdBu_r', list())
+    tfr._onselect(eclick, erelease, None, 'mean', None, 'RdBu_r')
     plt.close('all')