+++ mne/viz/topo.py
@@ -182,14 +182,15 @@ def _plot_topo_onpick(event, show_func=None, colorbar=False):
         raise err
 
 
-def _imshow_tfr(ax, ch_idx, tmin, tmax, vmin, vmax, ylim=None, tfr=None,
-                freq=None, vline=None, x_label=None, y_label=None,
+def _imshow_tfr(ax, ch_idx, tmin, tmax, vmin, vmax, onselect, ylim=None,
+                tfr=None, freq=None, vline=None, x_label=None, y_label=None,
                 colorbar=False, picker=True, cmap='RdBu_r', title=None):
     """ Aux function to show time-freq map on topo """
     import matplotlib.pyplot as plt
+    from matplotlib.widgets import RectangleSelector
     extent = (tmin, tmax, freq[0], freq[-1])
-    ax.imshow(tfr[ch_idx], extent=extent, aspect="auto", origin="lower",
-              vmin=vmin, vmax=vmax, picker=picker, cmap=cmap)
+    img = ax.imshow(tfr[ch_idx], extent=extent, aspect="auto", origin="lower",
+                    vmin=vmin, vmax=vmax, picker=picker, cmap=cmap)
     if isinstance(ax, plt.Axes):
         if x_label is not None:
             ax.set_xlabel(x_label)
@@ -201,9 +202,12 @@ def _imshow_tfr(ax, ch_idx, tmin, tmax, vmin, vmax, ylim=None, tfr=None,
         if y_label is not None:
             plt.ylabel(y_label)
     if colorbar:
-        plt.colorbar()
+        plt.colorbar(mappable=img)
     if title:
         plt.title(title)
+    if not isinstance(ax, plt.Axes):
+        ax = plt.gca()
+    ax.RS = RectangleSelector(ax, onselect=onselect)  # reference must be kept
 
 
 def _plot_timeseries(ax, ch_idx, tmin, tmax, vmin, vmax, ylim, data, color,